learn-python/PROJETOS/db-pipeline/requirements.txt
DateTime==5.5
et-xmlfile==1.1.0
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
zope.interface==7.0.3
